Only load extended models in OSS
